dvfs: tegra12: Don't enable GPU rail attaching to DVFS
Current Tegra12 GPU power-ungating code depends on DVFS initialization
bringing GPU rail ON when connecting to regulator. This results in
unnecessary power leakage while rail is ON but GPU is clamped, and
defers entry to system suspend state until GPU completes full power
ON/OFF cycle at least once. Otherwise, entering suspend with GPU rail
turned ON but GPU power ON state not finalized may damage the chip.

This commit fixed power-ungating and DVFS initialization on Tegra12
architecture, so that GPU rail is kept OFF when DVFS is attached, and
turned ON only when GPU is ungated and its power ON state is finalized.

Note that Tegra12 and Tegra13 architectures share the same ungating
code, so this fix is applied to Tegra13 as well.

Bug 200053091

Change-Id: Icb9cf5611764262571db19717ab37026e11bf89a
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/671349
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>