video: tegra: host: Rework unmanaged syncpts
Currently client managed syncpoints are given per-device, however,
in map-at-submit mode this will not work; Given that submits from
clients comes in random order, the client increments go wrong. In
addition, the current implementation allows releasing client
managed syncpoint when there are still submits on-the-fly, however,
as there is no reference counting, this may trigger hard-to-debug
issues where the same syncpoint is used for different purposes.

This patch reworks client managed syncpoints to work in similar
fashion as managed syncpoints: Per channel (map-at-open policy)
or per-channel-instance (map-at-submit policy). If we map
channels at channel open time, we release syncpoint when the
last user of the channel closes the channel. If we map channels
at submit, the client managed syncpoint is released when the
client dies.

Bug 1525996

Change-Id: I3549340501b4ad12cf3a0ff03e4984b75b0b986b
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/671396
GVS: Gerrit_Virtual_Submit