usb: phy: tegra: Update usb-pmc code sequence
Set run bit and enable port power of a port well
before we disable the PMC control on that port.

Otherwise there is a delay in shifting control
from PMC to USB, hence controller is missing the
chirp sequence when HS device is hot-plugged.
Thus HS devices are getting detected as FS
devices when we do a PMC hot-plug.

Bug 200061186

Change-Id: Ic1268cd0cc8d100e30fb1fd985446825d28e09e2
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: http://git-master/r/673972
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>