platform: tegra: isomgr: Fix usage of client_valid array
In some places client_valid[i] was being used to determine if the i'th
element of isoclient_info array was valid. This is incorrect. Instead
the 'name' field of the isoclient_info element should be usd to determine
if the element is valid.

In isomgr_init(), migrate the init of the completion field and the kref
refcount field into the section of code protected by the valid check.

Change-Id: I50bf76fa6df7ba2e2ba50a91d421c2b8f8e5801e
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/674168
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>