platform: tegra: isomgr: Fix usage of client_valid array
In some places client_valid[i] was being used to determine if the i'th
element of isoclient_info array was valid. This is incorrect. Instead
the 'name' field of the isoclient_info element should be usd to determine
if the element is valid.

In isomgr_init(), migrate the init of the completion field and the kref
refcount field into the section of code protected by the valid check.

Change-Id: I0c52b2e9de3f980d760bc37640af15b9f8f4760b
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/709120
Reviewed-by: Wen Yi <wyi@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>