NVAVP: Virtualization of AVP on T124
This change adds NVAVP client driver to the kernel tree, which
communicates with the NVAVP server on the RM partition.

The DT node sets the IVC queue used for communication with server
and AS details.

The client sets up the device /dev/tegra_avpchannel and accepts
IOCTL calls from video decode clients in usersapce. The client
takes care of connecting with the server and pass the data coming
from decode requests to the AVP server.

The address relocation of buffers passed are re-mapped to the AVP
AS and passed to the server.

The current solution does not support decoding encrypted video streams

Bug 200014729

Change-Id: I8e0154deb319fef4f38b8b4b744e712f32b02696
Signed-off-by: Deepak Kamurthy <dkamurthy@nvidia.com>
Reviewed-on: http://git-master/r/722304
(cherry picked from commit fd09c3243f2b0f20c21bc60cae7966be5dd26b6e)
Reviewed-on: http://git-master/r/724925
Tested-by: David Gorgen <dgorgen@nvidia.com>
Reviewed-by: Vladislav Buzov <vbuzov@nvidia.com>