iommu/tegra: modify SMMU_EX_PTBL_PAGE to avoid pfn
SMMU_MK_PDE converts physical address to SMMU PDE.
So, when extracting the page, first extract the physical address
of the page and convert to page by using phys_to_page. This makes the
macro independent of PAGE_SIZE.

If this is not done, locate_pte() gives us invalid virtual address
and kernel panic occurs when dereferencing it.

Bug 1526131

Change-Id: Id44031939df88eeb2593ed68442bef03f5383f0a
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/676518
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>