ata:ahci:fix "sleeping function called from invalid context"
This issue is because we were trying to sleep by holding a spinlock.
Spinlock are not needed here because there is no synchronization issue
in suspending/resuming the controller.
Spinlock were addded in earlier design of RTPM which are not required now
as the synchronization issues will be taken care by the pm core.

Bug 200072102

Change-Id: I7481ac2c6ddcdbfee9d9e500cd324e1b65609abb
Signed-off-by: Preetham Chandru R <pchandru@nvidia.com>
Reviewed-on: http://git-master/r/676521
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>
Tested-by: Venkat Moganty <vmoganty@nvidia.com>