thermal: nct72: Refactor NCT driver
By using Device Tree for defining and declaring a thermal zone,
thermal trip points infomation for the zone is maintained by of-thermal layer.

But current NCT driver copyies the data from of-thermal layer to the driver
during the initializing the driver and also whenever the data in of-thermal
layer is updated. This leads us a few bugs.

By using zone's accessor methods to get the data, we can avoid maintaining
duplicated data.

Bug 1600027
Bug 200008324

Change-Id: Ia4b315d95db60471c388ab88f2eb415708cd76c9
Signed-off-by: Hyungwoo Yang <hyungwooy@nvidia.com>
Reviewed-on: http://git-master/r/676719
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>