usb: gadget: xudc: toggle vbus when port stuck in inactive.
There is a small LFPS when port is in inactive state, it appears as
there is a glitch on the Tx lines of the host when host is removing termination.
Workaround for it is to start a timer when PLC event for Inactive entry is
received. If the port is still in INACTIVE after 100ms, and there was no
connect/disconnect event, toggle VBUS.

Bug 200040011

Change-Id: I63ed9c58771ea5df9cffffa02a0f259c0364e3fc
Signed-off-by: Hui Fu <hfu@nvidia.com>
Reviewed-on: http://git-master/r/678003
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Tested-by: Ashutosh Jha <ajha@nvidia.com>