arm: tn8: Enable CMU on HDMI
As per the HDMI standard, we should be defaulting to the defined
"limited" color range. In Tegra, we use CMU on dc1/HDMI to achieve
this.

In the userspace, NvCplService is responsible for setting the cmu_enable
node as per the user configuration. However, it actually picks up the
default setting based on what the kernel sets it.

Prior to this change, we were actually defaulting to cmu_enable being
off. This would actually cause us to send full range data with the
limited range avi infoframe and hence the "output is too dark" concern.
This change enables the cmu by default on HDMI instead.

Bug 1572199

Change-Id: I16cccba681f7208b3a5b24e069a211f277af1190
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Signed-off-by: Nitish <nrajguru@nvidia.com>
Reviewed-on: http://git-master/r/740531
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>