r8152: Disable random mac address
We want to explicitly hard fail if we've never fused a mac address. If
the mac address is not fused, there is a chance that the stability bits
for ethernet is also not fused.

In that scenario, with the random mac address, we'll get a user who
thinks his ethernet is fine (because of the random address), but
ethernet will quickly fail within minutes.

Bug 1611325

Change-Id: I79767fe05fa2fd40bf3fc5292d9e17a36e2d6151
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/707984
(cherry picked from commit 4860367055ac347a7c75540b4cc5f379ff99a8ce)
Reviewed-on: http://git-master/r/761451
Reviewed-by: Zheng Liu <zhliu@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>