thermal: tegra: soctherm: add HW PLLX offsetting
introduce HW PLLX offsetting for T210 platforms (tunable via
device tree)

In cases where Tsosc-Vmin is greater than Login-Vmin, and the
in a usecase where the logic-Vmin dips below the tsosc-Vmin,
the tsosc has to be disabled. For CPU, GPU tsosc when this happens
we have to fall back on pllx tsosc (sensor) and right before we do
so, we record the offset change between the pllx and CPU/GPU tsosc.

The above feature is supported by HW and this changhe configures the
HW appropriately to do so.

(Bear in mind, HW feature, once enabled, trumps the previously existing
SW offsetting feature, os even if the SW feature coexists in the code, it
doesn't come in the way of HW offsetting feature)

Bug 1613857

Change-Id: I6e9b6c6d3812668ba1a495e375bfcf3b44945e11
Signed-off-by: navneet kumar <navneetk@nvidia.com>
Reviewed-on: http://git-master/r/710464
Reviewed-by: Matthew Longnecker <mlongnecker@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>