mfd: max77620: add support to select sub-modules for similar chips
There are multiple devices which is compatible with max77620 with
little bit differences. Teh differences may be  with lower number
of sub modules inside the device.

Add support to list the available sub modules depends on device and
register those module. The selection of list will be done based on
compatibility and its device data.

bug 200077395

Change-Id: Ibaf68e0ea542b3b031f958d7e80095f823c1f51d
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/710681