regulator: max8973: Control EN pin gpio only when externally controlled
The new version of MAX77621 which is very similar to MAX8973 have the
SHTDN pin instead of EN pin and when this bit toggles, the content of
all register of MAX77621 get reset to their POR values.

Currently driver is toggling the STNDN pin to control the VOUT
enable/disable if valid GPIO is provided.

Make this GPIO to be toggle only if external control is enabled
otherwise set it  on enable state and toggle the EN bit. EN bit
will control the vout enable/disable and hence there will not be
any register reset.

bug 200084189

Change-Id: I609f7ad560c87c1f6698ed665bffe3343e29b926
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/712536