ARM: tegra: hawkeye: Control GPU regulator enable/disable through EN bit
ON Hawkeye, MAX77621 is used as GPU regulator and VOUT is enable/disable
by SHUTDN pin. Doing this, it reset the device configuration to default
and for enabling rail, it again required to restore the registers.

The enable/disable can be achieve by the EN bit through register access
and keep SHTDN pin as active all the time.

Remove the external control flag to control the enable/disable of the
regulator through EN bit only.

bug 200084189

Change-Id: I50009654a7d8bc7255b0308f606dde329e46428c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/712537