platform: tegra: la: Use clk_get_sys()
Instead of using clk_get(), use clk_get_sys() so we avoid the "wildcard"
EMC clock matching.

This patch is part of a series of patches to fix a camera underflow due
to missing emc clocks in IsoMgr. To correctly fix that bug we need to
remove the EMC wildcard match. As a result all usages of the EMC wildcard
matching need to be removed.

Bug 200083696

Change-Id: I7549cb5a771c2c07f5ffb662d7686a9c9beea132
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/713418
Reviewed-by: Hoang Pham <hopham@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>