tegra: dc: Avoid wildchar EMC clock matching
Use clk_get_sys("tegra_emc", "emc") to avoid the "wildcard"
EMC clock matching.

This patch is part of a series of patches to fix a camera underflow
due to missing emc clocks in IsoMgr. To correctly fix that bug
we need to remove the EMC wildcard match. As a result all
usages of the EMC wildcard matching need to be removed.

Bug 200083696

Change-Id: I77fefdf6b0e2417cb7189fe7a6cefc7bf8bd48bf
Signed-off-by: Hoang Pham <hopham@nvidia.com>
Reviewed-on: http://git-master/r/713953
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>