net: wireless: bcmdhd: Dynamically chip select
Create a dynamic selection between the BCM43241 chip and the BCM4354
chip configs. This emulates the cflags behaviors using global flags and
a single function that sets them up based on which chip is detected.

This definitely isn't pretty, but should mostly function, assuming that
the chip detection is run properly. If not, that can always be selected
by the board file or DT property.

Bug 200084777

Change-Id: Icc379cd17cdff6dc378c3654b7b4f685d18db440
Signed-off-by: Dan Willemsen <dwillemsen@nvidia.com>
Reviewed-on: http://git-master/r/720757
Reviewed-by: Om Prakash Singh <omp@nvidia.com>
Tested-by: Om Prakash Singh <omp@nvidia.com>
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>