video: Tegra: DP: fix unbal enable_irq at resume
To fix the "unbalanced enable for IRQ" stack dump when resuming from the LP0
suspend with no DP panel/monitor connected to the system. It happens when
the number of disable_irq() calling and enable_irq() calling do not match.
Also, this fix will update the connection status to recognize the unplugging
of DP panel/monitor during the suspend.
The stack dump takes 220mSec. This fix may help to reduce the LP0 suspend
resuming time.

bug 200034625
bug 200051954
bug 200070123
DEPENDS ON: <None>

Change-Id: I4cb9efe8c19ba46b94881c6a287db2f261b26937
Signed-off-by: Sungwook Kim <sungwookk@nvidia.com>
Reviewed-on: http://git-master/r/714561
Reviewed-by: Bob Johnston <bjohnston@nvidia.com>
Reviewed-by: Animesh Kishore <ankishore@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>