usb: xotg: fix sparse warnings
Fixes below warnings
warning: symbol 'session_supported' was not declared. Should it be static?
warning: symbol 'xotg_alloc_timer' was not declared. Should it be static?
warning: symbol 'xotg_test_timer_comp' was not declared. Should it be static?
warning: symbol 'b_srp_response_wait_timer' was not declared.
Should it be static?
warning: symbol 'xotg_timer_comp' was not declared. Should it be static?
warning: symbol 'b_srp_done_timer' was not declared. Should it be static?
warning: symbol 'xotg_init_timers' was not declared. Should it be static?
warning: symbol 'xotg_deinit_timers' was not declared. Should it be static?
warning: symbol 'xotg_enable_srp_detect' was not declared. Should it be static?
warning: symbol 'xotg_suspend_state' was not declared. Should it be static?
warning: symbol 'xotg_enable_gadget' was not declared. Should it be static?
warning: Using plain integer as NULL pointer
warning: Using plain integer as NULL pointer
warning: symbol 'xotg_suspend_host' was not declared. Should it be static?
warning: memset with byte count of -1
warning: memset with byte count of 0

Bug 200067946

Change-Id: Id0133539ec29e77a263f9dfde043483c27eaca3a
Signed-off-by: Ajay Gupta <ajayg@nvidia.com>
Reviewed-on: http://git-master/r/715314
Reviewed-by: Hui Fu <hfu@nvidia.com>
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>