video: tegra: dc: Fix ddc enable and disable
For hdmi2.0, tegra_hdmi_ddc_enable() and tegra_hdmi_ddc_disable()
functions need to be called before we can access the DDC (which itself
is refcounted).

However, seems like there is a bug in tegra_hdmi_edid_eld_setup(). If
the call to read the EDID fails (for example, bad EDID checksum), it
doesn't actually call the ddc_disable() in the error path. This meant
that any subsequent hot plug failed with actual i2c transaction failures
and even if we were plugged into a TV with a good EDID, we would
actually fail trying to read the EDID.

But, the code in tegra_hdmi_edid_eld_setup() blindly assumes that i2c
transactions will be happening and it calls the _enable and _disable
(rather than the possible virtual_edid read which might be picked up from
device tree). Instead of fixing this in the error path, this change
completely removes the enable and disable. They are already done as part
of the tegra_hdmi_ddc_i2c_xfer, so don't need to be re-done here.

Bug 1608233

Change-Id: I983604bacb253143899d32e1f41cf6fbd1a423f9
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/715627
(cherry picked from commit 83f97641ef9d6362f221851f6f5c23c46abbd082)
Reviewed-on: http://git-master/r/719971
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>