watchdog: tegra: timeout calculation
The watchdog timeout, requested by an application, should be passed
unchanged to the hardware access functions.
Currently, if the application requests N sec timeout, the first event
from the watchdog will come in N / 4, that might be unexpected FIQ
debugger popup.
With this change, f the application requests N seconds timeout, then
first event will come from watchdog in N seconds (it will be IRQ or FIQ,
if configured).
In N * 4 seconds, the system will be restarted, if watchdog is not
serviced.

Bug 1624103

Change-Id: I9ce2f48126e07a7c737a3fc248eb3292ec2e9ba1
Signed-off-by: dmitry pervushin <dpervushin@nvidia.com>
Reviewed-on: http://git-master/r/717824
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>