staging: android: fiq_debugger: avoid reduplicate dump
The FIQ source may stay asserted and in that case FIQ handler will
be entered again and again. This causes a reduplicate dump each
time the FIQ handler is entered, hence we adds a per-cpu flag to
avoid the reduplicate dump for a same core.

Bug 1511804

Change-Id: I27a76e86e8b6ffa7362094dbb389162e94ee3e38
Signed-off-by: Allen Yu <alleny@nvidia.com>
Reviewed-on: http://git-master/r/718498
Reviewed-by: Dmitry Pervushin <dpervushin@nvidia.com>
Tested-by: Dmitry Pervushin <dpervushin@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>