arm: tegra: vcm30t124: add extcon for ID/VBUS
Extcon framework can be used to emulate GPIOs/SFIOs that are not available
in a platform.  Add extcon-gpio entries in vcm30t124 so that derived platforms
can choose to allow SW control of OTG-ID and Gadget-VBUS status.

This also allows initial value for the extcon-gpio to be set by changing the
device tree entry.

Bug 1603256

Change-Id: I208a285e75ee9236ff56909d4fc6e4920f0bf118
Signed-off-by: Haribabu Narayanan <hnarayanan@nvidia.com>
Reviewed-on: http://git-master/r/724545
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>