regulator: max77620: add support to set FPS on shutdown
MAX77620 does not turn off those rails which are not FPS sourced when
PMIC is in OFF state. Some of the rails can not turn on during shutdown
also as it is need for system to execute code. On this case, set the FPS
of rail on shutdown so that it can be turn OFF when PMIC is on OFF state.

Add support to set the FPS on shutdown so that rail can be
turn OFF when PMIC is OFF>

bug 1624701

Change-Id: I5428d477f64270ee1af67bc8b102873db64de8b3
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
(cherry picked from commit fb9a3d19c55ae5f05dc0d526d7786b304e46468d)
Reviewed-on: http://git-master/r/729542