ARM: tegra: hawkeye: add dtb file for wifi SKU
On Hawkeye wifi SKU, the rail SD2 which is supply power to the
modem regulator on modem SKU can be disable on suspend. But it
can not be disable on modem SKU as modem need continuous power.

Add new DTB file to make this different behavior across different
SKU.

bug 1618960

Change-Id: I2b37dfdd421535a6b18a5fd4ecb1280cac901c86
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/729020
GVS: Gerrit_Virtual_Submit