tegra: dc: New trace event for latency measurement
Adds a new trace event to get the timestamp when
the dc gets unblocked on pre_syncpt_fence for a
particular flip.

This (sync_wt_ovr_syncpt_upd) alongwith the
existing trace events (flip_rcvd_syncpt_upd and
scanout_syncpt_upd) help us in measuring latency
from the point we receive a flip, the point at
which the frame is ready for scan-out and the
instant dc actually scans out the same frame.

Change-Id: Icc007a58f161570b7dab090bcc843cc8065a1bc6
Signed-off-by: Arun Swain <arswain@nvidia.com>
Reviewed-on: http://git-master/r/729291
Reviewed-by: Daniel Solomon <daniels@nvidia.com>
Tested-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>