arm: dts: change VDD_HV_SMPS9 to 2.9V
VDDIO_SDMMC3 is AP power rail for SD card and it is sourced by LDO9.
Currently the power topology is: Vbatt -> VDD_SYS -> SMPS9 (2.8V) ->
LDO9 (bypass mode with 2.8V or 1.8V). The concern is that for low speed
card, it is working on 2.8V. We found that there are voltage dropout at
load and it is about 50mV drop during a real usage of write/read.
Confirmed with the team that Vmin of VDDIO_SDMMC3 for Tegra K1 is now
2.8V, not 2.7V. So LDE decide that we have to raise VDD_HV_SMPS9 to
2.9V to meet the requirement.

bug 200099456

Change-Id: Ic10db7026ab3ab66e4eabb8df018a7e0556ac68f
Signed-off-by: Ian Chang <ianc@nvidia.com>
Reviewed-on: http://git-master/r/734537
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Tested-by: Laxman Dewangan <ldewangan@nvidia.com>