platform: tegra: iommu: cleanup t124 fixup table
clean up fixup entries for devices which do not need linear map
support and have iommus= property in their device tree nodes.

Although, VIC has a DT node, fixup entry can't be removed as with the
current logic, VI's swgroup is being assigned to vic and it can cause
issues. So, instead mark it as an error entry.

Bug 1362461

Change-Id: I3cbff2f945b85c2580aa76d4f718a05bdef43a0a
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/734759
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>