watchdog: tegra: get expiry count from DT
The total expiry count of Tegra WDTs is limited to HW design and depends
on skip configuration if supported. This change adds support to get the
expiry count through device tree. To be safe, we set the default expiry
count to 1, in case device tree doesn't provide the value.

Bug 200100035

Change-Id: I490dc21d3c1a217c3873a25cb60bba991bd80873
Signed-off-by: Allen Yu <alleny@nvidia.com>
Reviewed-on: http://git-master/r/736602
Reviewed-by: Mitch Luban <mluban@nvidia.com>