watchdog: tegra: revmap support for enable-on-init
To recover device from early hang before userspace watchdog daemon starts,
WDT needs to be enabled on init and a kernel WDT petting routine is preferable
since the time to userspace boot is non-deterministic and it is possible that
userspace may not provide a watchdog daemon. This change does following:

- Request a threaded irq (WDT_IRQ) if enable-on-init is provided in DT
- Reload WDT in the threaded interrupt context to probe CPU lockup
- Remove the IRQ handler once userspace watchdog daemon takes over WDT

For legacy chips like t124 and t132, WDT_IRQ is not available for WDT petting
since it's used for FIQ debugger. So they still expect userspace to boot and
start petting the watchdog.

Bug 200100035

Change-Id: I9d6345aa2a154dcb91adf0d42c044902dece4744
Signed-off-by: Allen Yu <alleny@nvidia.com>
Reviewed-on: http://git-master/r/736803
Reviewed-by: Mitch Luban <mluban@nvidia.com>