platform: tegra: MC: Add new LA API
Add a new latency allowance API to check if a possible LA value exists
for a given frequency/BW combination. The other API for this always
programs the LA if it turns out there is a possible LA value. This new
API is just a check and as such does not program the computed LA.

Bug 1637311

Change-Id: Ie71762f50d093d6482c7c088615dd198cf65dc84
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/737900
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>