NVAVP: Fix for incomplete thumbnails on Android
The earlier global structure was getting corrupted by the incessant
calls to SET CLOCK on AVP/VDE.
This led to the syncpt ID and value set to 0, which was used by
client for fence_wait and exited immediately. So we used to read
incomplete buffers to encode the thumbnail.

The fix uses local ivc message structures to overcome the issue.

The change also includes cleanup of client ctx and lock/unlock
across the pushbuffer ioctl.

Bug 200100975

Change-Id: I7dbb9afb7515934f5a5029158609383f0dd72082
Signed-off-by: Deepak Kamurthy <dkamurthy@nvidia.com>
Reviewed-on: http://git-master/r/738467
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aniket Bahadarpurkar <aniketb@nvidia.com>
Reviewed-by: Paresh Anandathirtha <paresha@nvidia.com>
Reviewed-by: Bob Johnston <bjohnston@nvidia.com>
Reviewed-by: Vladislav Buzov <vbuzov@nvidia.com>