regulator: max15569: Change compatible of max16989 to max16989x
Device MAX15569 is compatible with MAX16989 and driver of MAX15569
is used for MAX16989. There is many feature need to be added for
completion of MAX16989 and also it is require to have cleanups on
driver.

To reduce the number of changes and avoid the risk of break, create
new file for MAX16989 driver and change this driver file to support
max16989x as old driver.

bug 1627041

Change-Id: I1b90924592c43e4df724e6fe799f8ca90c742b77
signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
(cherry picked from commit 8bfb981e37af98935ef65960d8187269d45fd451)
Reviewed-on: http://git-master/r/751260
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>