video: tegra: Do not set mode in update_monspecs()
Asking for mode change with pclk = 0 will cause
division by zero because we don't check for this
case in tegra_dc_program_mode().

Right now the code assumes that setting mode with
pclk = 0 will disable the screen. However, we not
only not disable the screen, we crash the kernel
instead. So it should be safe to just remove the
code.

If we ever need to disable DC in !specs case (i.e.
monitor disconnected), it is better to explicitly
call appropriate disable function (e.g.
tegra_dc_disable()) from the call site that
actually needs this (e.g. hdmi state
machine), instead of playing tricks with weird
modes and making a lot of assumptions on how it
will be handled down the stack.

Bug 1631478

Change-Id: I7a10f637daf252197cc5707a4624f7d111c5b207
Signed-off-by: Anton Vorontsov <avorontsov@nvidia.com>
Reviewed-on: http://git-master/r/739405
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>